Move proxies to a separate module
